src/microservice-python/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.7
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.24.1
alembic==1.11.1
